+++ geozero-core/tests/polylabel.rs
@@ -2,7 +2,7 @@ use flatgeobuf::*;
 use geo::contains::Contains;
 use geo::Geometry;
 use geozero::error::Result;
-use geozero_core::geo::RustGeo;
+use geozero_core::geo_types::Geo;
 use polylabel::polylabel;
 use std::fs::File;
 use std::io::BufReader;
@@ -15,7 +15,7 @@ fn country_labels() -> Result<()> {
     while let Some(feature) = fgb.next()? {
         let props = feature.properties()?;
         let geometry = feature.geometry().unwrap();
-        let mut geo = RustGeo::new();
+        let mut geo = Geo::new();
         geometry.process(&mut geo, GeometryType::MultiPolygon)?;
         if let Geometry::MultiPolygon(mpoly) = geo.geometry() {
             if let Some(poly) = &mpoly.0.iter().next() {